meta-charset/meta-inside-script.html

<!DOCTYPE html>
<script>/* <meta charset=windows-1251> */</script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => assert_equals(document.charset, 'windows-1251'), "charset inside script should work");
</script>
